+++ refsanalysis.py
@@ -63,7 +63,7 @@ from forefst import (
 )
 
 PROG = "refsanalysis"
-VERSION = "1.11.2"
+VERSION = "1.11.3"